Inline code to instrument inside the test
