+++ instrumentation/nocode/src/test/config/nocode.yml
@@ -1,18 +1,18 @@
-- class: com.splunk.opentelemetry.instrumentation.nocode.SampleClass
+- class: com.splunk.opentelemetry.instrumentation.nocode.NocodeInstrumentationTest$SampleClass
   method: doSomething
   spanName: "this.getName()"
   attributes:
     - key: "details"
       value: this.getDetails()
 
-- class: com.splunk.opentelemetry.instrumentation.nocode.SampleClass
+- class: com.splunk.opentelemetry.instrumentation.nocode.NocodeInstrumentationTest$SampleClass
   method: throwException
   spanKind: SERVER
   attributes:
     - key: "five"
       value: param0.toString().substring(0)
 
-- class: com.splunk.opentelemetry.instrumentation.nocode.SampleClass
+- class: com.splunk.opentelemetry.instrumentation.nocode.NocodeInstrumentationTest$SampleClass
   method: doInvalidRule
   spanName: "this.thereIsNoSuchMethod()"
   spanKind: INVALID
