--- instrumentation/nocode/src/test/java/com/splunk/opentelemetry/instrumentation/nocode/SampleClass.java
@@ -1,22 +1,0 @@
-package com.splunk.opentelemetry.instrumentation.nocode;
-
-public class SampleClass {
-  public String getName() {
-    return "name";
-  }
-
-  public String getDetails() {
-    return "details";
-  }
-
-  public void throwException(int parameter) {
-    throw new UnsupportedOperationException("oh no");
-  }
-
-  public void doSomething() {
-  }
-
-  public void doInvalidRule() {
-  }
-
-}
